requirements.txt
asgiref==3.8.1
attrs==23.2.0
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crypto==1.4.1
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.6
django-filter==24.2
django-redis==5.4.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
drf_api_logger==1.1.15
idna==3.7
inflection==0.5.1
Naked==0.1.32
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==70.0.0
shellescape==3.8.1
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.1
webencodings==0.5.1
